Add Dependency Service class
